menu to profil with ajax ok

+++ app/ajax/handleMenuRole.php
@@ -4,13 +4,15 @@ $menuModule = array();
 $menuModule['error'] = true;
 $menuModule['message'] = "Opération échoué";
 $menuModule['data'] = array();
-$post = file_get_contents('php://input');
-if (!empty($_GET['action'])) {
+//Recuperetion des données send en input
+$post = json_decode(file_get_contents('php://input'), true) ?? $_POST;
+if (!empty($_GET['action'])) {//Vérification s'il y a une action à getté
     extract($_GET);
+    
     if ($action=='add') {
         if (!empty($post['id_profil']) && !empty($post['id_action'])) {
             $pdo = $GLOBALS['connexion'];
-            $sql = "INSERT INTO profil_has_action(id_profil, id_action) VALUES(? ?)";
+            $sql = "INSERT INTO profil_has_action(id_profil, id_action) VALUES(?, ?)";
             $req = $pdo->prepare($sql);
             $req->execute([$post['id_profil'],$post['id_action']]);
             $menuModule['error'] = false;
@@ -18,8 +20,8 @@ if (!empty($_GET['action'])) {
             $menuModule['data'] = $res;
         }
     }elseif ($action=='delete') {
-        //echo json_encode($post); die;
-        if (!empty($post)) {
+        
+        if (!empty($post['id_profil']) && !empty($post['id_action'])) {
             
             $pdo = $GLOBALS['connexion'];
             $sql = "DELETE FROM profil_has_action WHERE id_profil=? AND id_action=?";
@@ -27,7 +29,7 @@ if (!empty($_GET['action'])) {
             $req->execute([$post['id_profil'],$post['id_action']]);
             $menuModule['error'] = false;
             $menuModule['message'] = "Opération delete réussis";
-            $menuModule['data'] = $res;
+            $menuModule['data'] = array();
         }
     }
 }
